Changed double precisions to reals
After the modifications to the makefile, the
default state for the real numbers upon
compilation should be 64-bit. This makes it
more compatible with modern Fortran.

+++ exercises/se/ex1/particle.f90
@@ -5,7 +5,7 @@ MODULE particle
   TYPE particle3d
      TYPE(point3d)  :: p ! Position
      TYPE(vector3d) :: v ! Velocity
-     DOUBLE PRECISION :: m ! Mass
+     REAL :: m ! Mass
   END TYPE particle3d
